E News auth cheack

+++ app/Http/Controllers/AddController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Redirect;
 use App\Http\Requests;
@@ -20,36 +21,42 @@ class AddController extends Controller
     }
       // add store database by obydul date:7-8-16
     public function add_store(Request $request){
-        $validator = Validator::make($request->all(),[
-            'add_title'    =>'required',
-            'add_position' =>'required',
-            'back_link'    =>'required'
-        ]);
-        if($validator->fails()){
-            Session::flash('error', 'Something went wrong!');
-            return redirect::to("add-create-form")->withErrors($validator);
+        if(Auth::check()){
+
+            $validator = Validator::make($request->all(),[
+                'add_title'    =>'required',
+                'add_position' =>'required',
+                'back_link'    =>'required'
+            ]);
+            if($validator->fails()){
+                Session::flash('error', 'Something went wrong!');
+                return redirect::to("add-create-form")->withErrors($validator);
+            }
+            else{
+                $add_content = new AddModel();
+
+                //dd($add_content);
+                $image = Input::file('add_image');
+                $filename  = time() . '.' . $image->getClientOriginalExtension();
+                $path = public_path('image_folder/' . $filename);
+                Image::make($image->getRealPath())->resize(200, 200)->save($path);
+
+                $add_content->add_title = $request->get('add_title');
+                $add_content->add_image  = $filename ;
+                $add_content->position    = $request->get('add_position');
+                $add_content->back_link = $request->get('back_link');
+                $add_content->save();
+
+                Session::flash('success', 'Successfully Data Insert.');
+                return redirect::to('add-create-form');
+            }
         }
-        else{
-            $add_content = new AddModel();
+        else{}
 
-        //dd($add_content);
-            $image = Input::file('add_image');
-            $filename  = time() . '.' . $image->getClientOriginalExtension();
-            $path = public_path('image_folder/' . $filename);
-            Image::make($image->getRealPath())->resize(200, 200)->save($path);
-
-            $add_content->add_title = $request->get('add_title');
-            $add_content->add_image  = $filename ;
-            $add_content->position    = $request->get('add_position');
-            $add_content->back_link = $request->get('back_link');
-            $add_content->save();
-
-            Session::flash('success', 'Successfully Data Insert.');
-            return redirect::to('add-create-form');
-        }
 
     }
     public function add_update(Request $request,$id){
+        if(Auth::check()){
         $adds_data = AddModel::find($id);
 
         if (Input::hasFile('image_add')) {
@@ -77,21 +84,30 @@ class AddController extends Controller
         }
         Session::flash('success', 'Successfully updated.');
         return redirect::to('add-create-form');
+        }
+        else{}
 
     }
     // ads status by obydul dae:11-8-16
     public function  AdsStatus(Request $request,$id){
+        if(Auth::check()){
         $data         =  AddModel::find($id);
         $data->status = $request->get('ads_status');
         $data->save();
         Session::flash('success', 'Successfully  Inserted.');
         return redirect::to('add-create-form');
+        }
+        else{}
     }
 //adds delete by obydul date:10-8-16
     public function adds_delete($id){
+        if(Auth::check()){
         $sub_cat_delete= AddModel::find($id)->delete();
         return redirect::to('add-create-form');
+        }
+        else{}
     }
+
 
     //end class
 }
